burst: add new mode burst2 which may actaully be better
this one sends a burst and waits for the whole thing to complete, when it
does it sends a new burst. you can use PPS/BPS in this scenario as a
measure of bursts per second, and it probably hits the rx much more.

also added period size option to the original burst, but that made me
realise this approach is probably better

Signed-off-by: Dylan Yudaken <dylany@fb.com>