add support for maybe upcoming close on register
The idea is that if io_uring owns a file, that it can do neat things
like set socket nolock. However if your system does not allow you to
accept using accept_direct theres no way into it. So we add a flag to
io_uring_register_files to close the fd and take ownership of it (as
long as there are no other references). Still a WIP.

Signed-off-by: Dylan Yudaken <dylany@fb.com>